+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for the asynchronous "Info" dialogs found in KiCad 6.
+  Example: missing .kicad_wks (See INTI-CMNB/KiBot#264)
 
 ## [2.0.2] - 2022-08-26
 ### Fixed